requirements.txt
numpy==1.26.4
pyspark==3.5.0
pyarrow==17.0.0
pytest==9.0.2
azure-storage-file-datalake==12.12.0
pyodbc==4.0.39
ruff==0.6.9
python-dotenv==1.0.1
azure-mgmt-datafactory==7.0.0
azure-identity==1.19.0
